requirements.txt
streamlit==1.25.0
openai==1.91.0
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.25.0